refactor: folder service interface 생성
